Connection: detect installed systemd service or helm release
If the container is down for some reason we need to check if it
installed using other means.

In order to avoid import loops, the Connection object is moved up one
folder in shared.
