+++ mgradm/cmd/uninstall/uninstall.go
@@ -7,6 +7,7 @@ package uninstall
 import (
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
+	"github.com/uyuni-project/uyuni-tools/shared"
 	"github.com/uyuni-project/uyuni-tools/shared/kubernetes"
 	"github.com/uyuni-project/uyuni-tools/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
@@ -28,7 +29,7 @@ func NewCommand(globalFlags *types.GlobalFlags) *cobra.Command {
 				backend, _ = cmd.Flags().GetString("backend")
 			}
 
-			cnx := utils.NewConnection(backend, podman.ServerContainerName, kubernetes.ServerFilter)
+			cnx := shared.NewConnection(backend, podman.ServerContainerName, kubernetes.ServerFilter)
 			command, err := cnx.GetCommand()
 			if err != nil {
 				log.Fatal().Err(err).Msg("Failed to determine suitable backend")
