--- shared/utils/cp.go
@@ -1,83 +1,0 @@
-// SPDX-FileCopyrightText: 2023 SUSE LLC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package utils
-
-import (
-	"strings"
-
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-)
-
-// Copy transfers a file to or from the container.
-// Prefix one of src or dst parameters with `server:` to designate the path is in the container
-// user and group parameters are used to set the owner of a file transfered in the container.
-func Copy(cnx *Connection, src string, dst string, user string, group string) {
-	podName, err := cnx.GetPodName()
-	if err != nil {
-		log.Fatal().Err(err)
-	}
-	var commandArgs []string
-	extraArgs := []string{}
-	srcExpanded := strings.Replace(src, "server:", podName+":", 1)
-	dstExpanded := strings.Replace(dst, "server:", podName+":", 1)
-
-	command, err := cnx.GetCommand()
-	if err != nil {
-		log.Fatal().Err(err)
-	}
-
-	switch command {
-	case "podman-remote":
-		fallthrough
-	case "podman":
-		commandArgs = []string{"cp", srcExpanded, dstExpanded}
-	case "kubectl":
-		commandArgs = []string{"cp", "-c", "uyuni", srcExpanded, dstExpanded}
-		extraArgs = []string{"-c", "uyuni", "--"}
-	default:
-		log.Fatal().Msgf("Unknown container kind: %s", command)
-	}
-
-	RunCmdStdMapping(command, commandArgs...)
-
-	if user != "" && strings.HasPrefix(dst, "server:") {
-		execArgs := []string{"exec", podName}
-		execArgs = append(execArgs, extraArgs...)
-		owner := user
-		if group != "" {
-			owner = user + ":" + group
-		}
-		execArgs = append(execArgs, "chown", owner, strings.Replace(dst, "server:", "", 1))
-		RunCmdStdMapping(command, execArgs...)
-	}
-}
-
-func TestExistenceInPod(cnx *Connection, dstpath string) bool {
-	podName, err := cnx.GetPodName()
-	if err != nil {
-		log.Fatal().Err(err)
-	}
-	commandArgs := []string{"exec", podName}
-
-	command, err := cnx.GetCommand()
-	if err != nil {
-		log.Fatal().Err(err)
-	}
-
-	switch command {
-	case "podman":
-		commandArgs = append(commandArgs, "test", "-e", dstpath)
-	case "kubectl":
-		commandArgs = append(commandArgs, "-c", "uyuni", "test", "-e", dstpath)
-	default:
-		log.Fatal().Msgf("Unknown container kind: %s\n", command)
-	}
-
-	if _, err := RunCmdOutput(zerolog.DebugLevel, command, commandArgs...); err != nil {
-		return false
-	}
-	return true
-}
